+++ Backend/src/controllers/doctor.controller.js
@@ -12,6 +12,19 @@ exports.getAllDoctors = async (req, res) => {
 }
 
 exports.createDoctor = async (req, res) => {
+    if (!req.body.user_id) {
+        return baseResponse(res, false, 400, 'Missing user id', null);
+    }
+    if (!req.body.name) {
+        return baseResponse(res, false, 400, 'Missing name', null);
+    }
+    if (!req.body.specialization) {
+        return baseResponse(res, false, 400, 'Missing specialization', null);
+    }
+    if (!req.body.hospital_id) {
+        return baseResponse(res, false, 400, 'Missing hospital id', null);
+    }
+
     try {
         const doctor = await doctorRepository.createDoctor(req.body);
         baseResponse(res, true, 201, 'Doctor created', doctor);
@@ -21,6 +34,10 @@ exports.createDoctor = async (req, res) => {
 }
 
 exports.getDoctorById = async (req, res) => {
+    if (!req.params.id) {
+        return baseResponse(res, false, 400, 'Missing id');
+    }
+    
     try {
         const doctor = await doctorRepository.getDoctorById(req.params.id);
         if (!doctor) return baseResponse(res, false, 404, 'Doctor not found', null);
@@ -31,7 +48,22 @@ exports.getDoctorById = async (req, res) => {
 }
 
 exports.updateDoctor = async (req, res) => {
-    if (!req.body.id) return baseResponse(res, false, 400, 'Missing doctor id');
+    if (!req.body.id) {
+        return baseResponse(res, false, 400, 'Missing id');
+    }
+    if (!req.body.user_id) {
+        return baseResponse(res, false, 400, 'Missing user id', null);
+    }
+    if (!req.body.name) {
+        return baseResponse(res, false, 400, 'Missing name', null);
+    }
+    if (!req.body.specialization) {
+        return baseResponse(res, false, 400, 'Missing specialization', null);
+    }
+    if (!req.body.hospital_id) {
+        return baseResponse(res, false, 400, 'Missing hospital id', null);
+    }
+
     try {
         const doctor = await doctorRepository.updateDoctor(req.body);
         if (!doctor) return baseResponse(res, false, 404, 'Doctor not found', null);
@@ -42,6 +74,10 @@ exports.updateDoctor = async (req, res) => {
 }
 
 exports.deleteDoctor = async (req, res) => {
+    if (!req.params.id) {
+        return baseResponse(res, false, 400, 'Missing id');
+    }
+    
     try {
         const doctor = await doctorRepository.deleteDoctor(req.params.id);
         if (!doctor) return baseResponse(res, false, 404, 'Doctor not found', null);
