+++ Backend/src/controllers/hospital.controller.js
@@ -12,6 +12,13 @@ exports.getAllHospitals = async (req, res) => {
 }
 
 exports.createHospital = async (req, res) => {
+    if (!req.body.name) {
+        return baseResponse(res, false, 400, 'Missing name', null);
+    }
+    if (!req.body.address) {
+        return baseResponse(res, false, 400, 'Missing address', null);
+    }
+
     try {
         const hospital = await hospitalRepository.createHospital(req.body);
         baseResponse(res, true, 201, 'Hospital created', hospital);
@@ -21,6 +28,10 @@ exports.createHospital = async (req, res) => {
 }
 
 exports.getHospitalById = async (req, res) => {
+    if (!req.params.id) {
+        return baseResponse(res, false, 400, 'Missing id');
+    }
+
     try {
         const hospital = await hospitalRepository.getHospitalById(req.params.id);
         if (!hospital) return baseResponse(res, false, 404, 'Hospital not found', null);
@@ -31,7 +42,16 @@ exports.getHospitalById = async (req, res) => {
 }
 
 exports.updateHospital = async (req, res) => {
-    if (!req.body.id) return baseResponse(res, false, 400, 'Missing hospital id');
+    if (!req.body.id) {
+        return baseResponse(res, false, 400, 'Missing id');
+    }
+    if (!req.body.name) {
+        return baseResponse(res, false, 400, 'Missing name', null);
+    }
+    if (!req.body.address) {
+        return baseResponse(res, false, 400, 'Missing address', null);
+    }
+
     try {
         const hospital = await hospitalRepository.updateHospital(req.body);
         if (!hospital) return baseResponse(res, false, 404, 'Hospital not found', null);
@@ -42,6 +62,10 @@ exports.updateHospital = async (req, res) => {
 }
 
 exports.deleteHospital = async (req, res) => {
+    if (!req.params.id) {
+        return baseResponse(res, false, 400, 'Missing id');
+    }
+
     try {
         const hospital = await hospitalRepository.deleteHospital(req.params.id);
         if (!hospital) return baseResponse(res, false, 404, 'Hospital not found', null);
