+++ Backend/src/controllers/patient.controller.js
@@ -12,8 +12,17 @@ exports.getAllPatients = async (req, res) => {
 }
 
 exports.createPatient = async (req, res) => {
-    if (!req.body.user_id, !req.body.name || !req.body.age || !req.body.address) {
-        return baseResponse(res, false, 400, 'Missing patient data', null);
+    if (!req.body.user_id) {
+        return baseResponse(res, false, 400, 'Missing user id', null);
+    }
+    if (!req.body.name) {
+        return baseResponse(res, false, 400, 'Missing name', null);
+    }
+    if (!req.body.age) {
+        return baseResponse(res, false, 400, 'Missing age', null);
+    }
+    if (!req.body.address) {
+        return baseResponse(res, false, 400, 'Missing address', null);
     }
     
     try {
@@ -25,6 +34,10 @@ exports.createPatient = async (req, res) => {
 }
 
 exports.getPatientById = async (req, res) => {
+    if (!req.params.id) {
+        return baseResponse(res, false, 400, 'Missing id');
+    }
+
     try {
         const patient = await patientRepository.getPatientById(req.params.id);
         if (!patient) return baseResponse(res, false, 404, 'Patient not found', null);
@@ -35,10 +48,22 @@ exports.getPatientById = async (req, res) => {
 }
 
 exports.updatePatient = async (req, res) => {
-    if (!req.body.id) return baseResponse(res, false, 400, 'Missing patient id');
-    if (!req.body.user_id, !req.body.name || !req.body.age || !req.body.address) {
-        return baseResponse(res, false, 400, 'Missing patient data', null);
+    if (!req.body.id) {
+        return baseResponse(res, false, 400, 'Missing id');
     }
+    if (!req.body.user_id) {
+        return baseResponse(res, false, 400, 'Missing user id', null);
+    }
+    if (!req.body.name) {
+        return baseResponse(res, false, 400, 'Missing name', null);
+    }
+    if (!req.body.age) {
+        return baseResponse(res, false, 400, 'Missing age', null);
+    }
+    if (!req.body.address) {
+        return baseResponse(res, false, 400, 'Missing address', null);
+    }
+
     try {
         const patient = await patientRepository.updatePatient(req.body);
         if (!patient) return baseResponse(res, false, 404, 'Patient not found', null);
@@ -49,6 +74,10 @@ exports.updatePatient = async (req, res) => {
 }
 
 exports.deletePatient = async (req, res) => {
+    if (!req.params.id) {
+        return baseResponse(res, false, 400, 'Missing id');
+    }
+
     try {
         const patient = await patientRepository.deletePatient(req.params.id);
         if (!patient) return baseResponse(res, false, 404, 'Patient not found', null);
